fixed the online status with fsocket

+++ basefunc.php
@@ -205,7 +205,10 @@ function ServerOnline($link){
 	global $ServerPort;
 	$result = mysqli_query($link,"SELECT zoneid FROM point WHERE zoneid IS NOT NULL");
 	$count = @mysqli_num_rows($result);
-    $sockres = @FSockOpen($LanIP, $ServerPort, $errno, $errstr, 1);
+	if ($count > 0) {
+		return true;
+	}
+    $sockres = @FSockOpen("localhost", $ServerPort, $errno, $errstr, 1);
     if (!$sockres){
 		mysqli_query($link,"DELETE FROM online");
         $SRunning=false;
